+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - [FEATURE] Add form-loss-preventer service and use it on route mixin to enable multiple form loss preventions on one route
 - [FIX/BREAKING] Remove resetController handler in route mixin
+- [FEATURE] Add support for multiple forms on form route mixin
 
 # 0.2.9
 - [FIX] Setting calculated values only to writeable (!readonly) virtual properties